refactor(cli): drop app-local shared CLI shims
skillspp-cli still routed several commands through one-line local re-exports even
though the shared transport and UI modules already lived in cli-shared.
This points commands at the shared modules directly and moves duplicate
UI and background-runner tests into the shared package.

Constraint: Preserve app/package boundaries enforced by scripts/check-boundaries.sh
Rejected: Keep shim files for convenience | they hid ownership and duplicated tests
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Add new shared CLI helpers in cli-shared instead of re-exporting them from apps
Tested: vitest packages/cli-shared/src/runtime/background-runner-app-adapters.test.ts
Tested: vitest packages/cli-shared/src/ui/selection-step.test.ts
Not-tested: Full skillspp-cli command matrix
