--- apps/skillspp-cli/src/command-builder.ts
@@ -1,1 +1,0 @@
-export * from "@skillspp/cli-shared/command-builder";
