+++ apps/skillspp-cli/src/commands/update.ts
@@ -5,14 +5,14 @@ import { parsePolicyMode } from "../policy-mode";
 import {
   parseStandaloneCommand,
   type CliCommandContext,
-} from "../command-builder";
+} from "@skillspp/cli-shared/command-builder";
 import { runBackgroundTask } from "../runtime/background-runner";
-import { canUseInteractive } from "../interactive";
+import { canUseInteractive } from "@skillspp/cli-shared/interactive";
 import {
   type ManySelectionViewConfig,
   type SelectionKeyHint,
   runManySelectionStep,
-} from "../ui/selection-step";
+} from "@skillspp/cli-shared/ui/selection-step";
 import {
   completedStepsSection,
   failedStepsSection,
@@ -24,8 +24,11 @@ import {
   renderStaticScreen,
   showLoader,
   sourceSection,
-} from "../ui/screens";
-import { formatDriftChips, shortenHomePath } from "../ui/format";
+} from "@skillspp/cli-shared/ui/screens";
+import {
+  formatDriftChips,
+  shortenHomePath,
+} from "@skillspp/cli-shared/ui/format";
 import type { SelectionRow } from "@skillspp/core/agents";
 
 export type UpdateOptions = CheckOptions & {
