--- apps/skillspp-cli/src/interactive.ts
@@ -1,1 +1,0 @@
-export * from "@skillspp/cli-shared/interactive";
