--- apps/skillspp-cli/src/ui/colors.ts
@@ -1,1 +1,0 @@
-export * from "@skillspp/cli-shared/ui/colors";
