--- apps/skillspp-cli/src/ui/format.ts
@@ -1,1 +1,0 @@
-export * from "@skillspp/cli-shared/ui/format";
