--- apps/skillspp-cli/src/ui/logo.ts
@@ -1,1 +1,0 @@
-export * from "@skillspp/cli-shared/ui/logo";
