--- apps/skillspp-cli/src/ui/screens.tsx
@@ -1,1 +1,0 @@
-export * from "@skillspp/cli-shared/ui/screens";
