--- apps/skillspp-cli/src/ui/selection-step.tsx
@@ -1,1 +1,0 @@
-export * from "@skillspp/cli-shared/ui/selection-step";
